+++ packages/integrations/qqbot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @melandlabs/integrations-qqbot
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @melandlabs/env-config@0.4.0
+  - @melandlabs/ai@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes