+++ packages/loop/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @melandlabs/loop
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @melandlabs/env-config@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes